registration get 2 dup records

+++ app/src/main/java/com/o2sports/hxiao/o2sports_basketball/entity/Player.java
@@ -7,7 +7,7 @@ public class Player {
     public String id;
     public String name;
     public int position;
-    public double height;
-    public double weight;
+    public int height;
+    public int weight;
     public boolean gender;
 }
